Add `className` and `css` to `BaseProps`
`className` is a default DOM-level prop in React, and it feels wrong to
omit this since it's possible in vanilla JS to pass this.

`css` is exposed via styled-components, but only works with the Babel
plugin, which rewrites any element with a `css` prop and converts it
into a proper styled-component component. See
https://styled-components.com/docs/api#css-prop for more information.